Use int64 Index for native PDN n/nnz, not int32.
libdpn C API, Eigen SparseMatrix StorageIndex, and Python ctypes now
share int64_t indices so mesh size is not capped at 2^31-1. Event counts
still reject n_events > INT_MAX. SciPy fallback CSR may remain int32.

+++ engine/include/dpn/csr.hpp
@@ -7,7 +7,8 @@
 
 namespace dpn {
 
-using Index = int32_t;
+/* Mesh n and nnz are not capped at 2^31-1. Eigen sparse StorageIndex matches. */
+using Index = int64_t;
 
 struct Csr {
   Index nrows = 0;
